backend/requirements.txt
asgiref==3.11.0
Django==6.0.1
djangorestframework==3.14.0
django-cors-headers==4.6.0
sqlparse==0.5.5
psycopg2-binary==2.9.9
